requirements.txt
absl-py==1.0.0
blessings==1.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grad-cam==1.3.7
grpcio==1.46.1
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
munch==2.5.0
numpy
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pretrainedmodels==0.7.4
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.0
sklearn==0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
ttach==0.0.3
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
